fix(brandService.ts): moved BrandProfilePayload interface to types
